project/requirements.txt
django-model-utils==1.4.0
factory-boy==2.1.1
